fixed problem with post requests
Post requests didn't include the type parameter, so they would fail. For instance, the example in the readme for posting an image didn't work.

I'm not sure if this is the best way to solve the problem, but it makes it so I can submit post requests to the api.